+++ src/VirtoCommerce.Platform.Caching/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using StackExchange.Redis;
 using VirtoCommerce.Platform.Core.Caching;
 using VirtoCommerce.Platform.Redis;
 
@@ -21,6 +22,10 @@ namespace VirtoCommerce.Platform.Caching
                 services.AddOptions<RedisCachingOptions>().Bind(configuration.GetSection("Caching:Redis")).ValidateDataAnnotations();
 
                 services.AddSingleton<IPlatformMemoryCache, RedisPlatformMemoryCache>();
+
+                var redis = ConnectionMultiplexer.Connect(redisConnectionString);
+                services.AddSingleton<IConnectionMultiplexer>(redis);
+                services.AddSingleton(redis.GetSubscriber());
             }
             else
             {
