+++ src/VirtoCommerce.Platform.Web/Extensions/ServiceCollectionExtensions.cs
@@ -1,71 +1,32 @@
-using System;
-using System.Linq;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Infrastructure;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using VirtoCommerce.Platform.Core.Common;
-using VirtoCommerce.Platform.Core.Modularity;
-using VirtoCommerce.Platform.Modules.External;
-using VirtoCommerce.Platform.Web.Infrastructure;
 
 namespace VirtoCommerce.Platform.Modules
 {
     public static class ServiceCollectionExtensions
     {
-        public static IServiceCollection AddModules(this IServiceCollection services, IMvcBuilder mvcBuilder, Action<LocalStorageModuleCatalogOptions> setupAction = null)
+        public static IServiceCollection ConfigureServer(this IServiceCollection services)
         {
-            services.AddSingleton(services);
-
-            services.AddSingleton<IModuleInitializer, ModuleInitializer>();
-            // Cannot inject IHostingEnvironment to LoadContextAssemblyResolver as IsDevelopment() is an extension method (means static) and cannot be mocked by Moq in tests
-            services.AddSingleton<IAssemblyResolver, LoadContextAssemblyResolver>(provider =>
-                new LoadContextAssemblyResolver(provider.GetService<ILogger<LoadContextAssemblyResolver>>(), provider.GetService<IWebHostEnvironment>().IsDevelopment()));
-            services.AddSingleton<IModuleManager, ModuleManager>();
-            services.AddSingleton<ILocalModuleCatalog, LocalStorageModuleCatalog>();
-            services.AddSingleton<IModuleCatalog>(provider => provider.GetService<ILocalModuleCatalog>());
-
-            if (setupAction != null)
+            // Enable synchronous IO if using Kestrel:
+            services.Configure<KestrelServerOptions>(options =>
             {
-                services.Configure(setupAction);
-            }
-            var providerSnapshot = services.BuildServiceProvider();
+                options.AllowSynchronousIO = true;
+            });
 
-            var manager = providerSnapshot.GetRequiredService<IModuleManager>();
-            var moduleCatalog = providerSnapshot.GetRequiredService<ILocalModuleCatalog>();
-
-            manager.Run();
-            // Ensure all modules are loaded
-            foreach (var module in moduleCatalog.Modules.OfType<ManifestModuleInfo>().Where(x => x.State == ModuleState.NotStarted).ToArray())
+            // Enable synchronous IO if using IIS:
+            services.Configure<IISServerOptions>(options =>
             {
-                manager.LoadModule(module.ModuleName);
+                options.AllowSynchronousIO = true;
+            });
 
-                // VP-2190: No need to add parts for modules with laoding errors - it could cause an exception
-                if (module.Assembly != null && module.Errors.IsNullOrEmpty())
-                {
-                    // Register API controller from modules
-                    mvcBuilder.AddApplicationPart(module.Assembly);
-                }
-            }
-
-            services.AddSingleton(moduleCatalog);
-            return services;
-        }
-
-        public static IServiceCollection AddExternalModules(this IServiceCollection services, Action<ExternalModuleCatalogOptions> setupAction = null)
-        {
-            services.AddSingleton<IExternalModulesClient, ExternalModulesClient>();
-            services.AddSingleton<IExternalModuleCatalog, ExternalModuleCatalog>();
-            services.AddSingleton<IModuleInstaller, ModuleInstaller>();
-
-            if (setupAction != null)
+            services.Configure<CookiePolicyOptions>(options =>
             {
-                services.Configure(setupAction);
-            }
-
-            services.AddSingleton<IPlatformRestarter, ProcessPlatformRestarter>();
+                // This lambda determines whether user consent for non-essential cookies is needed for a given request.
+                options.CheckConsentNeeded = context => true;
+                options.MinimumSameSitePolicy = SameSiteMode.None;
+            });
 
             return services;
         }
